# backend/docs/api-contracts/openapi-contracts.yaml
openapi: 3.0.0
info:
  title: Contracts Microservice API
  description: Manages contract lifecycle and contract state transitions
  version: 1.0.0
  contact:
    name: Systems Engineering Team

servers:
  - url: http://localhost:8082/api/contracts
    description: Local development server

security:
  - BearerAuth: []

paths:
  /create:
    post:
      summary: Create a new contract
      description: Register a new contract (ADMINISTRADOR or FUNCIONARIO roles)
      operationId: createContract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Contract created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '400':
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Supplier not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /list:
    get:
      summary: List all contracts
      description: Retrieve all contracts with optional filtering and pagination
      operationId: listContracts
      parameters:
        - name: status
          in: query
          description: Filter by contract status (SIGNED, ACTIVE, SUSPENDED, COMPLETED, CANCELED)
          required: false
          schema:
            type: string
            enum: [SIGNED, ACTIVE, SUSPENDED, COMPLETED, CANCELED]
        - name: supplierId
          in: query
          description: Filter by supplier UUID
          required: false
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search by contract number or description
          required: false
          schema:
            type: string
            maxLength: 100
        - name: limit
          in: query
          description: Maximum number of results
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of contracts
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{contractId}:
    get:
      summary: Get contract by ID
      description: Retrieve detailed information about a specific contract
      operationId: getContract
      parameters:
        - name: contractId
          in: path
          description: UUID of the contract
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contract details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '400':
          description: Invalid contract ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /update/{contractId}:
    put:
      summary: Update contract details
      description: Update contract information (ADMINISTRADOR or FUNCIONARIO, only SIGNED contracts)
      operationId: updateContract
      parameters:
        - name: contractId
          in: path
          description: UUID of the contract to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Contract updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '400':
          description: Invalid request or contract ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - can only update SIGNED contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /transition/{contractId}:
    patch:
      summary: Transition contract state
      description: Change contract state according to business rules (status transitions)
      operationId: transitionContractState
      parameters:
        - name: contractId
          in: path
          description: UUID of the contract
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
      responses:
        '200':
          description: Contract state transitioned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '400':
          description: Invalid transition or contract state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /history/{contractId}:
    get:
      summary: Get contract state history
      description: Retrieve all state transitions for a contract (audit trail)
      operationId: getContractHistory
      parameters:
        - name: contractId
          in: path
          description: UUID of the contract
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contract state history
          content:
            application/json:
              schema:
                type: object
                properties:
                  contractId:
                    type: string
                    format: uuid
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/StateTransitionRecord'
        '400':
          description: Invalid contract ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  schemas:
    CreateContractRequest:
      type: object
      required:
        - contractNumber
        - supplierId
        - description
        - startDate
        - endDate
        - amount
      properties:
        contractNumber:
          type: string
          maxLength: 50
          description: Unique contract identifier/number
        supplierId:
          type: string
          format: uuid
          description: UUID of the supplier
        description:
          type: string
          maxLength: 1000
          description: Contract description and scope
        startDate:
          type: string
          format: date
          description: Contract start date (YYYY-MM-DD)
        endDate:
          type: string
          format: date
          description: Contract end date (YYYY-MM-DD)
        amount:
          type: number
          format: double
          minimum: 0
          description: Contract total amount in currency
        notes:
          type: string
          maxLength: 500
          nullable: true
          description: Additional notes (optional)

    UpdateContractRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
          description: Contract description (if updating)
        amount:
          type: number
          format: double
          minimum: 0
          description: Contract amount (if updating)
        notes:
          type: string
          maxLength: 500
          nullable: true
          description: Additional notes (if updating)

    TransitionRequest:
      type: object
      required:
        - newStatus
      properties:
        newStatus:
          type: string
          enum: [ACTIVE, SUSPENDED, COMPLETED, CANCELED]
          description: New contract status (from current SIGNED or previous state)
        reason:
          type: string
          maxLength: 500
          nullable: true
          description: Reason for state transition (optional but recommended)

    ContractResponse:
      type: object
      properties:
        contractId:
          type: string
          format: uuid
          description: Unique contract identifier
        contractNumber:
          type: string
          description: Contract reference number
        supplierId:
          type: string
          format: uuid
          description: Supplier UUID
        description:
          type: string
          description: Contract description
        status:
          type: string
          enum: [SIGNED, ACTIVE, SUSPENDED, COMPLETED, CANCELED]
          description: Current contract status
        startDate:
          type: string
          format: date
          description: Start date (YYYY-MM-DD)
        endDate:
          type: string
          format: date
          description: End date (YYYY-MM-DD)
        amount:
          type: number
          format: double
          description: Total amount
        notes:
          type: string
          nullable: true
          description: Additional notes
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (UTC)

    StateTransitionRecord:
      type: object
      properties:
        fromStatus:
          type: string
          enum: [SIGNED, ACTIVE, SUSPENDED, COMPLETED, CANCELED]
          description: Previous status
        toStatus:
          type: string
          enum: [SIGNED, ACTIVE, SUSPENDED, COMPLETED, CANCELED]
          description: New status
        transitionReason:
          type: string
          nullable: true
          description: Reason for transition
        transitionBy:
          type: string
          format: uuid
          description: User UUID who performed the transition
        transitionAt:
          type: string
          format: date-time
          description: Transition timestamp (UTC)

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp (UTC)