# backend/proveedor-service/proveedor-service/src/main/java/com/distribuidos/proveedor_service/dto/ErrorResponse.java
package com.distribuidos.proveedor_service.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;
import java.util.Map;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ErrorResponse {
    
    private String codigo;
    private String mensaje;
    private List<Map<String, String>> detalles;
    private LocalDateTime timestamp;
    private String path;
}